+++ packages/asc-ui/src/components/Pagination/index.ts
@@ -1,0 +1,2 @@
+export * as PaginationStyle from './PaginationStyle'
+export { default as Pagination } from './Pagination'