+++ packages/asc-ui/src/components/TextField/TextFieldStyle.ts
@@ -1,14 +1,20 @@
 import styled from 'styled-components'
-import ButtonStyle from '../Button/ButtonStyle'
+import Button from '../Button'
+import { themeSpacing } from '../../utils'
 
 const TextFieldStyle = styled.div`
   position: relative;
+`
 
-  & > ${ButtonStyle} {
-    position: absolute;
-    top: 5px;
-    right: 5px;
-  }
+export const InputWrapper = styled.div`
+  position: relative;
+`
+
+export const CloseButton = styled(Button)`
+  position: absolute;
+  right: ${themeSpacing(2)};
+  top: 50%;
+  transform: translateY(-50%);
 `
 
 export default TextFieldStyle