+++ stories/src/ui/SearchBarToggle.stories.tsx
@@ -2,16 +2,13 @@ import { SearchBarToggle } from '@amsterdam/asc-ui'
 import { action } from '@storybook/addon-actions'
 import { useState } from 'react'
 
-const ControlledSearchBarToggle = ({ hasBackDrop = false }) => {
+export const ControlledSearchBarToggle = ({ hasBackDrop = false }) => {
   const [open, setOpen] = useState(false)
   return (
     <SearchBarToggle
       align="left"
       open={open}
       onOpen={setOpen}
-      onWatchValue={(value) => {
-        action(`input value changed: ${value}`)
-      }}
       onSubmit={() => {
         action('submit')
       }}